LayoutTests/fast/js/dfg-uint32-to-number-in-middle-of-copy-propagation-expected.txt

Tests that UInt32ToNumber and OSR exit are aware of copy propagation and correctly recover both versions of a variable that was subject to a UInt32ToNumber cast.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS foo(-1) is [-1, 4294967295]
PASS successfullyParsed is true

TEST COMPLETE
